MINOR: Handle expandIsr in PartitionLockTest and ensure read threads not blocked on write (#7973)
Fixed an intermittent failure in the test that was caused by a mocked instance not handling expandIsr. Also updated the test to use the offset from the batch to trigger expandIsr more often. With this change, the test may try to acquire write lock while appends are blocked on read lock, so separated out the test using write lock to make it safer.

Reviewers: Jason Gustafson <jason@confluent.io>