KAFKA-9605; Do not attempt to abort batches when txn manager is in fatal error (#8177)
We detected a bug in soak where the producer batches shall be failed in sender loop before the produce response callback. This shall trigger an illegal state exception on the producer batch as it is already aborted. 

The impact is not severe since sender is on its own thread but should be fixed to avoid unnecessary critical exception.

Reviewers: Bob Barrett <bob.barrett@confluent.io>, Jason Gustafson <jason@confluent.io>